Move DBSCAN to clustering
